Refactor bits of Sheet Util
