+++ tests/__init__.py
@@ -1,0 +1,4 @@
+import logging as log
+import os
+
+log.basicConfig(level=os.environ.get("LOGLEVEL"), format='%(asctime)s %(levelname)s %(message)s')
